Move related RDLUpdater to db discovery modules
